Move to rely on jackson module
