+++ diff/build.gradle
@@ -4,7 +4,7 @@ plugins {
 
 dependencies {
     api project(":api")
-    implementation project(":logstash")
+    implementation project(":jackson")
 
     // https://github.com/flipkart-incubator/zjsonpatch/
     // jackson dependency on 2.12.1
@@ -13,7 +13,6 @@ dependencies {
     implementation 'com.flipkart.zjsonpatch:zjsonpatch:0.4.12'
 
     // implementation 'de.danielbechler:java-object-diff:0.95'
-
 
     testImplementation project(":logger")
     testImplementation 'ch.qos.logback:logback-classic:1.2.10'
