+++ diff/src/main/java/com/tersesystems/echopraxia/diff/DiffObjectMapper.java
@@ -1,0 +1,13 @@
+package com.tersesystems.echopraxia.diff;
+
+import com.fasterxml.jackson.databind.ObjectMapper;
+
+public class DiffObjectMapper {
+  private DiffObjectMapper() {}
+
+  public static final ObjectMapper mapper = new ObjectMapper();
+
+  static {
+    mapper.findAndRegisterModules();
+  }
+}
